+++ src/main/java/de/nierhain/danger/Danger.java
@@ -4,14 +4,11 @@ import de.nierhain.danger.proxy.CommonProxy;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.SidedProxy;
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLServerStartingEvent;
+import net.minecraftforge.fml.common.event.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-@Mod(modid = Danger.MODID, name = Danger.NAME, version = Danger.VERSION)
+@Mod(modid = Danger.MODID, name = Danger.NAME, version = Danger.VERSION, certificateFingerprint = "3b024834bcbc5c9badb97afe9e26ed9ab7182a7d")
 public class Danger {
     public static final String MODID = "danger";
     public static final String NAME = "Danger Lies Ahead";
@@ -24,6 +21,12 @@ public class Danger {
 
     @Mod.Instance(MODID)
     public static Danger instance;
+
+    @EventHandler
+    public void onFingerprintViolation(FMLFingerprintViolationEvent event) {
+
+        logger.warn("Invalid fingerprint detected! The file " + event.getSource().getName() + " may have been tampered with. This version will NOT be supported by the author and you should get a valid copy ASAP!");
+    }
 
     @EventHandler
     public void preInit(FMLPreInitializationEvent event){