+++ src/main/java/de/nierhain/danger/network/PacketLevelUpToClient.java
@@ -1,6 +1,5 @@
 package de.nierhain.danger.network;
 
-import com.sun.xml.internal.bind.v2.runtime.reflect.Lister;
 import de.nierhain.danger.event.EventNotifyPlayer;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.client.Minecraft;
@@ -8,7 +7,6 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
-import scala.collection.parallel.ParIterableLike;
 
 public class PacketLevelUpToClient implements IMessage {