+++ Changelog.md
@@ -1,9 +1,10 @@
 ## 1.0
 
 - rewrote attribute system
+- rewrote configuration
 - removed skillpoints from level capability
 - finished first version of skill gui
-- added level up notification / particles / effects
+- added level up notification
 - fixed mobs getting levels by relogging into the world
 - sync levels between client and server
 - added german translation